requirements.txt
# requirements.txt
boto3==1.18.78
python-dotenv==0.19.1
